Change gradle external dependency management
This commit changes gradle to look for mathlibra.zip in the current
directory instead of downloading it from the build server. This approach
provides better flexibility and ensures that the build-server built
mathlibra.zip is not used, as mathlibraJ has to be built using the
same compiler as mathlibra.zip. Thus if the build server's files are
used a compiler version mismatch can occur.